test: e2e forester (#1615)

+++ sdk-libs/photon-api/src/models/context.rs
@@ -13,11 +13,11 @@ use crate::models;
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Context {
     #[serde(rename = "slot")]
-    pub slot: i32,
+    pub slot: u64,
 }
 
 impl Context {
-    pub fn new(slot: i32) -> Context {
+    pub fn new(slot: u64) -> Context {
         Context { slot }
     }
 }
